Check condition

+++ keep-n-from-N/exec.sh
@@ -15,6 +15,11 @@ ls -lh  ./files > ./output/gen_${timestamp}_sort_before.txt
 num_dump_files_to_keep=$2
 num_dump_files=$( ls -1 ./files | grep -E "db-[0-9]+\.dump" | wc -l )
 num_dump_files_to_delete=$(( $num_dump_files - $num_dump_files_to_keep ))
+if [ $num_dump_files_to_keep -ge $num_dump_files ] 
+	then
+		echo "No dump files to delete: Keep=$num_dump_files_to_keep. Total=$num_dump_files"
+		exit 0
+fi
 ls -1 ./files | grep -E "db-[0-9]+\.dump" | sort | head -n $num_dump_files_to_delete | xargs -I {} rm ./files/{}
 ls -lh  ./files > ./output/gen_${timestamp}_sort_after.txt
